+++ BitwardenKit/Core/Platform/Models/Domain/EnvironmentURLsTests.swift
@@ -1,7 +1,6 @@
-import BitwardenKit
 import XCTest
 
-@testable import BitwardenShared
+@testable import BitwardenKit
 
 class EnvironmentURLsTests: BitwardenTestCase {
     // MARK: Tests
